refactor: extract training spot list models
